[Tests] add shared lifecycle tests; move identical tests to shared
