+++ packages/enzyme-test-suite/test/ReactWrapper-spec.jsx
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 import { expect } from 'chai';
 import sinon from 'sinon-sandbox';
 import wrap from 'mocha-wrap';
-import isEqual from 'lodash.isequal';
 import {
   mount,
   ReactWrapper,
@@ -28,7 +27,6 @@ import {
   lazy,
   memo,
   Profiler,
-  PureComponent,
   Suspense,
 } from './_helpers/react-compat';
 import {
@@ -38,6 +36,7 @@ import {
 } from './_helpers';
 import getLoadedLazyComponent from './_helpers/getLoadedLazyComponent';
 import describeMethods from './_helpers/describeMethods';
+import describeLifecycles from './_helpers/describeLifecycles';
 import describeHooks from './_helpers/describeHooks';
 import {
   is,
@@ -1103,6 +1102,17 @@ describeWithDOM('mount', () => {
     'unmount',
     'wrap',
   );
+  describeLifecycles(
+    { Wrap, Wrapper },
+    'componentDidCatch',
+    'componentDidMount',
+    'componentDidUpdate',
+    'componentWillUnmount',
+    'getDerivedStateFromError',
+    'getDerivedStateFromProps',
+    'getSnapshotBeforeUpdate',
+    'misc',
+  );
   describeHooks(
     { Wrap, Wrapper },
     'useCallback',
@@ -1668,1990 +1678,6 @@ describeWithDOM('mount', () => {
     });
   });
 
-  describe('lifecycle methods', () => {
-    const errorToThrow = new EvalError('threw an error!');
-    // in React 16.0 - 16.2 and 16.9+, and some older nodes, the actual error thrown isn't reported.
-    const reactError = new Error('An error was thrown inside one of your components, but React doesn\'t know what it was. This is likely due to browser flakiness. React does its best to preserve the "Pause on exceptions" behavior of the DevTools, which requires some DEV-mode only tricks. It\'s possible that these don\'t work in your browser. Try triggering the error in production mode, or switching to a modern browser. If you suspect that this is actually an issue with React, please file an issue.');
-    const properErrorMessage = (error) => error instanceof Error && (
-      error.message === errorToThrow.message
-      || error.message === reactError.message
-    );
-
-    describeIf(is('>= 16.3'), 'getDerivedStateFromProps', () => {
-      let spy;
-
-      beforeEach(() => {
-        spy = sinon.spy();
-      });
-
-      class Spy extends React.Component {
-        constructor(...args) {
-          super(...args);
-          this.state = { state: true }; // eslint-disable-line react/no-unused-state
-          spy('constructor');
-        }
-
-        shouldComponentUpdate(nextProps, nextState, nextContext) {
-          spy('shouldComponentUpdate', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState,
-            prevContext: this.context,
-            nextContext,
-          });
-          return true;
-        }
-
-        componentWillUpdate(nextProps, nextState, nextContext) {
-          spy('componentWillUpdate', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState,
-            prevContext: this.context,
-            nextContext,
-          });
-        }
-
-        componentDidUpdate(prevProps, prevState, prevContext) {
-          spy('componentDidUpdate', {
-            prevProps,
-            nextProps: this.props,
-            prevState,
-            nextState: this.state,
-            prevContext,
-            nextContext: this.context,
-          });
-        }
-
-        render() {
-          spy('render');
-          return null;
-        }
-      }
-
-      class CWRP extends Spy {
-        componentWillReceiveProps(nextProps, nextContext) {
-          spy('componentWillReceiveProps', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState: this.state,
-            prevContext: this.context,
-            nextContext,
-          });
-        }
-      }
-
-      class U_CWRP extends Spy {
-        UNSAFE_componentWillReceiveProps(nextProps) { // eslint-disable-line camelcase
-          spy('UNSAFE_componentWillReceiveProps', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState: this.state,
-            prevContext: this.context,
-            nextContext: undefined,
-          });
-        }
-      }
-
-      class GDSFP extends Spy {
-        static getDerivedStateFromProps(props, state) {
-          spy('getDerivedStateFromProps', { props, state });
-          return {};
-        }
-      }
-
-      it('calls cWRP when expected', () => {
-        const prevProps = { a: 1 };
-        const wrapper = mount(<CWRP {...prevProps} />);
-        expect(spy.args).to.deep.equal([
-          ['constructor'],
-          ['render'],
-        ]);
-        spy.resetHistory();
-
-        const foo = {};
-        const props = { foo };
-        const {
-          context: prevContext,
-          context: nextContext,
-          state: prevState,
-          state: nextState,
-        } = wrapper.instance();
-
-        wrapper.setProps(props);
-        const nextProps = { ...prevProps, ...props };
-
-        const data = {
-          prevProps,
-          nextProps,
-          prevState,
-          nextState,
-          prevContext,
-          nextContext,
-        };
-        expect(spy.args).to.deep.equal([
-          ['componentWillReceiveProps', data],
-          ['shouldComponentUpdate', data],
-          ['componentWillUpdate', data],
-          ['render'],
-          ['componentDidUpdate', {
-            ...data,
-            prevContext: is('>= 16') ? undefined : prevContext,
-          }],
-        ]);
-      });
-
-      it('calls UNSAFE_cWRP when expected', () => {
-        const prevProps = { a: 1 };
-        // eslint-disable-next-line react/jsx-pascal-case
-        const wrapper = mount(<U_CWRP {...prevProps} />);
-        expect(spy.args).to.deep.equal([
-          ['constructor'],
-          ['render'],
-        ]);
-        spy.resetHistory();
-
-        const foo = {};
-        const props = { foo };
-        const {
-          context: prevContext,
-          context: nextContext,
-          state: prevState,
-          state: nextState,
-        } = wrapper.instance();
-
-        wrapper.setProps(props);
-        const nextProps = { ...prevProps, ...props };
-
-        const data = {
-          prevProps,
-          nextProps,
-          prevState,
-          nextState,
-          prevContext,
-          nextContext,
-        };
-        expect(spy.args).to.deep.equal([
-          ['UNSAFE_componentWillReceiveProps', {
-            ...data,
-            nextContext: is('>= 16') ? undefined : nextContext,
-          }],
-          ['shouldComponentUpdate', data],
-          ['componentWillUpdate', data],
-          ['render'],
-          ['componentDidUpdate', {
-            ...data,
-            prevContext: is('>= 16') ? undefined : prevContext,
-          }],
-        ]);
-      });
-
-      it('calls gDSFP when expected', () => {
-        const prevProps = { a: 1 };
-        const state = { state: true };
-        const wrapper = mount(<GDSFP {...prevProps} />);
-        expect(spy.args).to.deep.equal([
-          ['constructor'],
-          ['getDerivedStateFromProps', {
-            props: prevProps,
-            state,
-          }],
-          ['render'],
-        ]);
-        spy.resetHistory();
-
-        const foo = {};
-        const props = { foo };
-        const {
-          context: prevContext,
-          context: nextContext,
-          state: prevState,
-          state: nextState,
-        } = wrapper.instance();
-
-        wrapper.setProps(props);
-        const nextProps = { ...prevProps, ...props };
-
-        const data = {
-          prevProps,
-          nextProps,
-          prevState,
-          nextState,
-          prevContext,
-          nextContext,
-        };
-        expect(spy.args).to.deep.equal([
-          ['getDerivedStateFromProps', {
-            props: nextProps,
-            state: nextState,
-          }],
-          ['shouldComponentUpdate', data],
-          ['render'],
-          ['componentDidUpdate', {
-            ...data,
-            prevContext: is('>= 16') ? undefined : prevContext,
-          }],
-        ]);
-      });
-
-      it('cDU’s nextState differs from `this.state` when gDSFP returns new state', () => {
-        class SimpleComponent extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = { value: props.value };
-          }
-
-          static getDerivedStateFromProps(props, state) {
-            return props.value === state.value ? null : { value: props.value };
-          }
-
-          shouldComponentUpdate(nextProps, nextState) {
-            // eslint-disable-next-line react/destructuring-assignment
-            return nextState.value !== this.state.value;
-          }
-
-          render() {
-            const { value } = this.state;
-            return (<input value={value} />);
-          }
-        }
-        const wrapper = mount(<SimpleComponent value="initial" />);
-
-        expect(wrapper.find('input').prop('value')).to.equal('initial');
-
-        wrapper.setProps({ value: 'updated' });
-
-        expect(wrapper.find('input').prop('value')).to.equal('updated');
-      });
-    });
-
-    describeIf(is('>= 16'), 'componentDidCatch', () => {
-      describe('errors inside an error boundary', () => {
-        const hasFragments = is('>= 16.2');
-        const MaybeFragment = hasFragments ? Fragment : 'main';
-
-        function Thrower({ throws }) {
-          if (throws) {
-            throw errorToThrow;
-          }
-          return null;
-        }
-
-        class ErrorBoundary extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              throws: false,
-              didThrow: false,
-            };
-          }
-
-          componentDidCatch(error, info) {
-            const { spy } = this.props;
-            spy(error, info);
-            this.setState({
-              throws: false,
-              didThrow: true,
-            });
-          }
-
-          render() {
-            const {
-              didThrow,
-              throws,
-            } = this.state;
-
-            return (
-              <div>
-                <MaybeFragment>
-                  <span>
-                    <Thrower throws={throws} />
-                    <div>
-                      {didThrow ? 'HasThrown' : 'HasNotThrown'}
-                    </div>
-                  </span>
-                </MaybeFragment>
-              </div>
-            );
-          }
-        }
-
-        function ErrorSFC(props) {
-          return <ErrorBoundary {...props} />;
-        }
-
-        describe('Thrower', () => {
-          it('does not throw when `throws` is `false`', () => {
-            expect(() => mount(<Thrower throws={false} />)).not.to.throw();
-          });
-
-          it('throws when `throws` is `true`', () => {
-            expect(() => mount(<Thrower throws />)).to.throw();
-            try {
-              mount(<Thrower throws />);
-              expect(true).to.equal(false, 'this line should not be reached');
-            } catch (e) {
-              expect(e).to.satisfy(properErrorMessage);
-            }
-          });
-        });
-
-        it('catches a simulated error', () => {
-          const spy = sinon.spy();
-          const wrapper = mount(<ErrorBoundary spy={spy} />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError, info]] = spy.args;
-          expect(actualError).to.equal(errorToThrow);
-          expect(info).to.deep.equal({
-            componentStack: `
-    in Thrower (created by ErrorBoundary)
-    in span (created by ErrorBoundary)${hasFragments ? '' : `
-    in main (created by ErrorBoundary)`}
-    in div (created by ErrorBoundary)
-    in ErrorBoundary (created by WrapperComponent)
-    in WrapperComponent`,
-          });
-        });
-
-        it('rerenders on a simulated error', () => {
-          const wrapper = mount(<ErrorBoundary spy={sinon.stub()} />);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(1);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(0);
-        });
-
-        it('rerenders on a simulated error with an SFC root', () => {
-          const wrapper = mount(<ErrorSFC spy={sinon.stub()} />);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(1);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(0);
-        });
-
-        it('catches errors during render', () => {
-          const spy = sinon.spy();
-          const wrapper = mount(<ErrorBoundary spy={spy} />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          wrapper.setState({ throws: true });
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError, info]] = spy.args;
-          expect(actualError).to.satisfy(properErrorMessage);
-          expect(info).to.deep.equal({
-            componentStack: `
-    in Thrower (created by ErrorBoundary)
-    in span (created by ErrorBoundary)${hasFragments ? '' : `
-    in main (created by ErrorBoundary)`}
-    in div (created by ErrorBoundary)
-    in ErrorBoundary (created by WrapperComponent)
-    in WrapperComponent`,
-          });
-        });
-
-        it('works when the root is an SFC', () => {
-          const spy = sinon.spy();
-          const wrapper = mount(<ErrorSFC spy={spy} />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          wrapper.find(ErrorBoundary).setState({ throws: true });
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError, info]] = spy.args;
-          expect(actualError).to.satisfy(properErrorMessage);
-          expect(info).to.deep.equal({
-            componentStack: `
-    in Thrower (created by ErrorBoundary)
-    in span (created by ErrorBoundary)${hasFragments ? '' : `
-    in main (created by ErrorBoundary)`}
-    in div (created by ErrorBoundary)
-    in ErrorBoundary (created by ErrorSFC)
-    in ErrorSFC (created by WrapperComponent)
-    in WrapperComponent`,
-          });
-        });
-      });
-    });
-
-    describeIf(is('>= 16.6'), 'getDerivedStateFromError', () => {
-      describe('errors inside an error boundary', () => {
-        function Thrower({ throws }) {
-          if (throws) {
-            throw errorToThrow;
-          }
-          return null;
-        }
-
-        function getErrorBoundary() {
-          return class ErrorBoundary extends React.Component {
-            static getDerivedStateFromError() {
-              return {
-                throws: false,
-                didThrow: true,
-              };
-            }
-
-            constructor(props) {
-              super(props);
-              this.state = {
-                throws: false,
-                didThrow: false,
-              };
-            }
-
-            render() {
-              const {
-                didThrow,
-                throws,
-              } = this.state;
-
-              return (
-                <div>
-                  <Fragment>
-                    <span>
-                      <Thrower throws={throws} />
-                      <div>
-                        {didThrow ? 'HasThrown' : 'HasNotThrown'}
-                      </div>
-                    </span>
-                  </Fragment>
-                </div>
-              );
-            }
-          };
-        }
-
-        function ErrorSFC({ component }) {
-          return component();
-        }
-
-        describe('Thrower', () => {
-          it('does not throw when `throws` is `false`', () => {
-            expect(() => mount(<Thrower throws={false} />)).not.to.throw();
-          });
-
-          it('throws when `throws` is `true`', () => {
-            expect(() => mount(<Thrower throws />)).to.throw();
-            try {
-              mount(<Thrower throws />);
-              expect(true).to.equal(false, 'this line should not be reached');
-            } catch (error) {
-              expect(error).to.satisfy(properErrorMessage);
-            }
-          });
-        });
-
-        it('catches a simulated error', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const spy = sinon.spy(ErrorBoundary, 'getDerivedStateFromError');
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError]] = spy.args;
-          expect(actualError).to.equal(errorToThrow);
-        });
-
-        it('rerenders on a simulated error', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(1);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(0);
-        });
-
-        it('rerenders on a simulated error with an SFC root', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const wrapper = mount(<ErrorSFC component={() => <ErrorBoundary />} />);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(1);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(0);
-        });
-
-        it('catches errors during render', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const spy = sinon.spy(ErrorBoundary, 'getDerivedStateFromError');
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          wrapper.setState({ throws: true });
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError]] = spy.args;
-          expect(actualError).to.satisfy(properErrorMessage);
-        });
-
-        it('works when the root is an SFC', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const spy = sinon.spy(ErrorBoundary, 'getDerivedStateFromError');
-          const wrapper = mount(<ErrorSFC component={() => <ErrorBoundary />} />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          wrapper.find(ErrorBoundary).setState({ throws: true });
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError]] = spy.args;
-          expect(actualError).to.satisfy(properErrorMessage);
-        });
-      });
-    });
-
-    describeIf(is('>= 16.6'), 'getDerivedStateFromError and componentDidCatch combined', () => {
-      const expectedInfo = {
-        componentStack: `
-    in Thrower (created by ErrorBoundary)
-    in div (created by ErrorBoundary)
-    in ErrorBoundary (created by WrapperComponent)
-    in WrapperComponent`,
-      };
-
-      function Thrower({ throws }) {
-        if (throws) {
-          throw errorToThrow;
-        }
-        return null;
-      }
-
-      describe('errors inside error boundary when getDerivedStateFromProps returns update', () => {
-        let lifecycleSpy;
-        let stateSpy;
-
-        beforeEach(() => {
-          lifecycleSpy = sinon.spy();
-          stateSpy = sinon.spy();
-        });
-
-        class ErrorBoundary extends React.Component {
-          static getDerivedStateFromError(error) {
-            lifecycleSpy('getDerivedStateFromError', error);
-            return {
-              didThrow: true,
-              throws: false,
-            };
-          }
-
-          constructor(props) {
-            super(props);
-            this.state = {
-              didThrow: false,
-              throws: false,
-            };
-
-            lifecycleSpy('constructor');
-          }
-
-          componentDidCatch(error, info) {
-            lifecycleSpy('componentDidCatch', error, info);
-            stateSpy({ ...this.state });
-          }
-
-          render() {
-            lifecycleSpy('render');
-
-            const {
-              throws,
-            } = this.state;
-
-            return (
-              <div>
-                <Thrower throws={throws} />
-              </div>
-            );
-          }
-        }
-
-        it('calls getDerivedStateFromError first and then componentDidCatch', () => {
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(lifecycleSpy).to.have.property('callCount', 2);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          expect(stateSpy).to.have.property('callCount', 0);
-
-          lifecycleSpy.resetHistory();
-
-          wrapper.setState({ throws: true });
-
-          expect(lifecycleSpy).to.have.property('callCount', 4);
-          const [first, second, third, fourth] = lifecycleSpy.args;
-          expect(first).to.deep.equal(['render']);
-          expect(second).to.satisfy(([name, error, ...rest]) => name === 'getDerivedStateFromError'
-              && properErrorMessage(error)
-              && rest.length === 0);
-          expect(third).to.deep.equal(['render']);
-          expect(fourth).to.satisfy(([name, error, info]) => name === 'componentDidCatch' && properErrorMessage(error) && isEqual(info, expectedInfo));
-
-          expect(stateSpy).to.have.property('callCount', 1);
-          expect(stateSpy.args).to.deep.equal([
-            [{
-              throws: false,
-              didThrow: true,
-            }],
-          ]);
-        });
-
-        it('calls getDerivedStateFromError first and then componentDidCatch on a simulated error', () => {
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(lifecycleSpy).to.have.property('callCount', 2);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          expect(stateSpy).to.have.property('callCount', 0);
-
-          lifecycleSpy.resetHistory();
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(lifecycleSpy).to.have.property('callCount', 3);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['getDerivedStateFromError', errorToThrow],
-            ['render'],
-            ['componentDidCatch', errorToThrow, expectedInfo],
-          ]);
-
-          expect(stateSpy).to.have.property('callCount', 1);
-          expect(stateSpy.args).to.deep.equal([
-            [{
-              throws: false,
-              didThrow: true,
-            }],
-          ]);
-        });
-      });
-
-      describe('errors inside error boundary when getDerivedStateFromError does not return update', () => {
-        let spy;
-
-        beforeEach(() => {
-          spy = sinon.spy();
-        });
-
-        class ErrorBoundary extends React.Component {
-          static getDerivedStateFromError(error) {
-            spy('getDerivedStateFromError', error);
-            return null;
-          }
-
-          constructor(props) {
-            super(props);
-            this.state = {
-              didThrow: false,
-              throws: false,
-            };
-
-            spy('constructor');
-          }
-
-          componentDidCatch(error, info) {
-            spy('componentDidCatch', error, info);
-
-            this.setState({
-              didThrow: true,
-              throws: false,
-            });
-          }
-
-          render() {
-            spy('render');
-
-            const {
-              throws,
-              didThrow,
-            } = this.state;
-
-            return (
-              <div>
-                <Thrower throws={throws} />
-                <div>
-                  {didThrow ? 'HasThrown' : 'HasNotThrown'}
-                </div>
-              </div>
-            );
-          }
-        }
-
-        it('renders again without calling componentDidCatch and then fails', () => {
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 2);
-          expect(spy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          spy.resetHistory();
-
-          try {
-            wrapper.setState({ throws: true });
-            expect('should never get here').to.equal(false);
-          } catch (e) {
-            expect(e).to.satisfy(properErrorMessage);
-          }
-
-          expect(spy).to.have.property('callCount', 3);
-          const [first, second, third] = spy.args;
-          expect(first).to.deep.equal(['render']);
-          expect(second).to.satisfy(([name, arg, ...rest]) => name === 'getDerivedStateFromError'
-              && properErrorMessage(arg)
-              && rest.length === 0);
-          expect(third).to.deep.equal(['render']);
-        });
-
-        it('renders again on simulated error', () => {
-          const wrapper = mount(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 2);
-          expect(spy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          spy.resetHistory();
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(spy).to.have.property('callCount', 3);
-          expect(spy.args).to.deep.equal([
-            ['getDerivedStateFromError', errorToThrow],
-            ['componentDidCatch', errorToThrow, expectedInfo],
-            ['render'],
-          ]);
-        });
-      });
-    });
-
-    context('mounting phase', () => {
-      it('calls componentWillMount and componentDidMount', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          componentWillMount() {
-            spy('componentWillMount');
-          }
-
-          componentDidMount() {
-            spy('componentDidMount');
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-        mount(<Foo />);
-        expect(spy.args).to.deep.equal([
-          ['componentWillMount'],
-          ['render'],
-          ['componentDidMount'],
-        ]);
-      });
-
-      it('is batching updates', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillMount() {
-            /* eslint-disable react/destructuring-assignment */
-            this.setState({ count: this.state.count + 1 });
-            this.setState({ count: this.state.count + 1 });
-            /* eslint-enable react/destructuring-assignment */
-          }
-
-          componentDidMount() {
-            /* eslint-disable react/destructuring-assignment */
-            this.setState({ count: this.state.count + 1 });
-            this.setState({ count: this.state.count + 1 });
-            /* eslint-enable react/destructuring-assignment */
-          }
-
-          render() {
-            spy();
-            const { count } = this.state;
-            return <div>{count}</div>;
-          }
-        }
-        const result = mount(<Foo />);
-        expect(result.state('count')).to.equal(2);
-        expect(spy).to.have.property('callCount', 2);
-      });
-    });
-
-    context('updating props', () => {
-      it('calls shouldComponentUpdate, componentWillUpdate, and componentDidUpdate', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              foo: 'state',
-            };
-          }
-
-          componentWillReceiveProps(nextProps, nextContext) {
-            spy('componentWillReceiveProps', this.props, nextProps, nextContext);
-          }
-
-          shouldComponentUpdate(nextProps, nextState, nextContext) {
-            spy('shouldComponentUpdate', this.props, nextProps, this.state, nextState, nextContext);
-            return true;
-          }
-
-          componentWillUpdate(nextProps, nextState, nextContext) {
-            spy('componentWillUpdate', this.props, nextProps, this.state, nextState, nextContext);
-          }
-
-          componentDidUpdate(prevProps, prevState, prevContext) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, prevContext);
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-
-        const wrapper = mount(
-          <Foo foo="bar" />,
-          {
-            context: { foo: 'context' },
-          },
-        );
-        wrapper.setProps({ foo: 'baz' });
-        wrapper.setProps({ foo: 'bax' });
-        expect(spy.args).to.deep.equal([
-          [
-            'render',
-          ],
-          [
-            'componentWillReceiveProps',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'context' }, // this will be fixed
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'state' }, { foo: 'state' },
-            is('>= 16') ? undefined : { foo: 'context' },
-          ],
-          [
-            'componentWillReceiveProps',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'context' },
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'state' }, { foo: 'state' },
-            is('>= 16') ? undefined : { foo: 'context' },
-          ],
-        ]);
-      });
-
-      it('calls componentWillReceiveProps, shouldComponentUpdate, componentWillUpdate and componentDidUpdate with merged props', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          componentWillReceiveProps(nextProps) {
-            spy('componentWillReceiveProps', this.props, nextProps);
-          }
-
-          shouldComponentUpdate(nextProps) {
-            spy('shouldComponentUpdate', this.props, nextProps);
-            return true;
-          }
-
-          componentWillUpdate(nextProps) {
-            spy('componentWillUpdate', this.props, nextProps);
-          }
-
-          componentDidUpdate(prevProps) {
-            spy('componentDidUpdate', prevProps, this.props);
-          }
-
-          render() {
-            return (
-              <div />
-            );
-          }
-        }
-
-        const wrapper = mount(<Foo a="a" b="b" />);
-
-        wrapper.setProps({ b: 'c', d: 'e' });
-
-        expect(spy.args).to.deep.equal([
-          [
-            'componentWillReceiveProps',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-          [
-            'shouldComponentUpdate',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-          [
-            'componentWillUpdate',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-          [
-            'componentDidUpdate',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-        ]);
-      });
-
-      it('cancels rendering when Component returns false in shouldComponentUpdate', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          componentWillReceiveProps() {
-            spy('componentWillReceiveProps');
-          }
-
-          shouldComponentUpdate() {
-            spy('shouldComponentUpdate');
-            return false;
-          }
-
-          componentWillUpdate() {
-            spy('componentWillUpdate');
-          }
-
-          componentDidUpdate() {
-            spy('componentDidUpdate');
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-
-        const wrapper = mount(<Foo foo="bar" />);
-        expect(wrapper.instance().props.foo).to.equal('bar');
-        wrapper.setProps({ foo: 'baz' });
-        expect(wrapper.instance().props.foo).to.equal('baz');
-        wrapper.setProps({ foo: 'bax' });
-        expect(wrapper.instance().props.foo).to.equal('bax');
-        expect(spy.args).to.deep.equal([
-          ['render'],
-          ['componentWillReceiveProps'],
-          ['shouldComponentUpdate'],
-          ['componentWillReceiveProps'],
-          ['shouldComponentUpdate'],
-        ]);
-      });
-
-      it('does not provoke another renders to call setState in componentWillReceiveProps', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillReceiveProps() {
-            /* eslint-disable react/destructuring-assignment */
-            this.setState({ count: this.state.count + 1 });
-            this.setState({ count: this.state.count + 1 });
-            /* eslint-enable react/destructuring-assignment */
-          }
-
-          render() {
-            spy();
-            const { foo } = this.props;
-            return <div>{foo}</div>;
-          }
-        }
-        const result = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setProps({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 2);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      it('provokes an another render to call setState twice in componentWillUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { foo } = this.props;
-            return <div>{foo}</div>;
-          }
-        }
-        const result = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setProps({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      it('provokes an another render to call setState twice in componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentDidUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/no-did-update-set-state, react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/no-did-update-set-state, react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { foo } = this.props;
-            return <div>{foo}</div>;
-          }
-        }
-        const result = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setProps({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-    });
-
-    context('updating state', () => {
-      it('calls shouldComponentUpdate, componentWillUpdate and componentDidUpdate', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              foo: 'bar',
-            };
-          }
-
-          shouldComponentUpdate(nextProps, nextState, nextContext) {
-            spy('shouldComponentUpdate', this.props, nextProps, this.state, nextState, nextContext);
-            return true;
-          }
-
-          componentWillUpdate(nextProps, nextState, nextContext) {
-            spy('componentWillUpdate', this.props, nextProps, this.state, nextState, nextContext);
-          }
-
-          componentDidUpdate(prevProps, prevState, prevContext) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, prevContext);
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-
-        const wrapper = mount(
-          <Foo foo="props" />,
-          {
-            context: { foo: 'context' },
-          },
-        );
-        wrapper.setState({ foo: 'baz' });
-        const expected = [
-          [
-            'render',
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'context' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'context' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'bar' }, { foo: 'baz' },
-            is('>= 16') ? undefined : { foo: 'context' },
-          ],
-        ];
-        expect(spy.args).to.deep.equal(expected);
-      });
-
-      it('cancels rendering when Component returns false in shouldComponentUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'bar',
-            };
-          }
-
-          shouldComponentUpdate() {
-            spy('shouldComponentUpdate');
-            return false;
-          }
-
-          componentWillUpdate() {
-            spy('componentWillUpdate');
-          }
-
-          componentDidUpdate() {
-            spy('componentDidUpdate');
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const wrapper = mount(<Foo />);
-        expect(wrapper.instance().state.foo).to.equal('bar');
-        wrapper.setState({ foo: 'baz' });
-        expect(wrapper.instance().state.foo).to.equal('baz');
-        expect(spy.args).to.deep.equal([['render'], ['shouldComponentUpdate']]);
-      });
-
-      it('provokes an another render to call setState twice in componentWillUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              name: 'foo',
-              count: 0,
-            };
-          }
-
-          componentWillUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setState({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      it('provokes an another render to call setState twice in componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              name: 'foo',
-              count: 0,
-            };
-          }
-
-          componentDidUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/no-did-update-set-state, react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/no-did-update-set-state, react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setState({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-    });
-
-    context('updating context', () => {
-      it('calls shouldComponentUpdate, componentWillUpdate and componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              foo: 'state',
-            };
-          }
-
-          shouldComponentUpdate(nextProps, nextState, nextContext) {
-            spy('shouldComponentUpdate', this.props, nextProps, this.state, nextState, nextContext);
-            return true;
-          }
-
-          componentWillUpdate(nextProps, nextState, nextContext) {
-            spy('componentWillUpdate', this.props, nextProps, this.state, nextState, nextContext);
-          }
-
-          componentDidUpdate(prevProps, prevState, prevContext) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, prevContext);
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-        const wrapper = mount(
-          <Foo foo="props" />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        expect(wrapper.instance().context.foo).to.equal('bar');
-        wrapper.setContext({ foo: 'baz' });
-        expect(wrapper.instance().context.foo).to.equal('baz');
-        expect(spy.args).to.deep.equal([
-          [
-            'render',
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'baz' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'baz' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'state' }, { foo: 'state' },
-            is('>= 16') ? undefined : { foo: 'bar' },
-          ],
-        ]);
-      });
-
-      it('cancels rendering when Component returns false in shouldComponentUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          shouldComponentUpdate() {
-            spy('shouldComponentUpdate');
-            return false;
-          }
-
-          componentWillUpdate() {
-            spy('componentWillUpdate');
-          }
-
-          componentDidUpdate() {
-            spy('componentDidUpdate');
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-        const wrapper = mount(
-          <Foo />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        wrapper.setContext({ foo: 'baz' });
-        expect(spy.args).to.deep.equal([['render'], ['shouldComponentUpdate']]);
-      });
-
-      it('provokes an another render to call setState twice in componentWillUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = mount(
-          <Foo />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        expect(spy).to.have.property('callCount', 1);
-        result.setContext({ foo: 'baz' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      it('provokes an another render to call setState twice in componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentDidUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/no-did-update-set-state, react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/no-did-update-set-state, react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = mount(
-          <Foo />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        expect(spy).to.have.property('callCount', 1);
-        result.setContext({ foo: 'baz' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-    });
-
-    context('unmounting phase', () => {
-      it('calls componentWillUnmount', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          componentWillUnmount() {
-            spy();
-          }
-
-          render() {
-            return <div>foo</div>;
-          }
-        }
-        const wrapper = mount(<Foo />);
-        wrapper.unmount();
-        expect(spy).to.have.property('callCount', 1);
-      });
-    });
-
-    context('component instance', () => {
-      it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          componentDidUpdate() {}
-
-          onChange() {
-            this.setState({ foo: 'onChange update' });
-          }
-
-          render() {
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-        const wrapper = mount(<Foo />);
-        wrapper.setState({ foo: 'wrapper setState update' });
-        expect(wrapper.state('foo')).to.equal('wrapper setState update');
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.instance().onChange();
-        expect(wrapper.state('foo')).to.equal('onChange update');
-        expect(spy).to.have.property('callCount', 2);
-      });
-
-      it('calls `componentDidUpdate` when component’s `setState` is called through a bound method', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-            this.onChange = this.onChange.bind(this);
-          }
-
-          componentDidUpdate() {}
-
-          onChange() {
-            this.setState({ foo: 'onChange update' });
-          }
-
-          render() {
-            const { foo } = this.state;
-            return (
-              <div>
-                {foo}
-                <button type="button" onClick={this.onChange}>click</button>
-              </div>
-            );
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-        const wrapper = mount(<Foo />);
-        wrapper.find('button').prop('onClick')();
-        expect(wrapper.state('foo')).to.equal('onChange update');
-        expect(spy).to.have.property('callCount', 1);
-      });
-
-      it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-            this.update = () => this.setState({ foo: 'update' });
-          }
-
-          componentDidMount() {
-            this.update();
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-        const wrapper = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        expect(wrapper.state('foo')).to.equal('update');
-      });
-
-      it('does not call `componentDidMount` twice when a child component is created', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          componentDidMount() {}
-
-          render() {
-            const { foo } = this.state;
-            return (
-              <div>
-                <button type="button" onClick={() => this.setState({ foo: 'update2' })}>
-                  click
-                </button>
-                {foo}
-              </div>
-            );
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidMount');
-
-        const wrapper = mount(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.find('button').prop('onClick')();
-        expect(spy).to.have.property('callCount', 1);
-      });
-    });
-
-    describeIf(is('>= 15.3'), 'PureComponent', () => {
-      it('does not update when state and props did not change', () => {
-        class Foo extends PureComponent {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-        const wrapper = mount(<Foo id={1} />);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-      });
-
-      class Test extends PureComponent {
-        constructor(...args) {
-          super(...args);
-
-          this.state = { a: { b: { c: 1 } } };
-        }
-
-        componentDidUpdate() {
-          const { onUpdate } = this.props;
-          onUpdate();
-        }
-
-        setDeepEqualState() {
-          this.setState({ a: { b: { c: 1 } } });
-        }
-
-        setDeepDifferentState() {
-          this.setState({ a: { b: { c: 2 } } });
-        }
-
-        render() {
-          const { a: { b: { c } } } = this.state;
-          return <div>{c}</div>;
-        }
-      }
-
-      it('rerenders on setState when new state is !==, but deeply equal to existing state', () => {
-        const updateSpy = sinon.spy();
-        const wrapper = mount(<Test onUpdate={updateSpy} />);
-        wrapper.instance().setDeepEqualState();
-        expect(updateSpy).to.have.property('callCount', 1);
-      });
-
-      it('rerenders when setState is called with an object that doesnt have deep equality', () => {
-        const updateSpy = sinon.spy();
-        const wrapper = mount(<Test onUpdate={updateSpy} />);
-        wrapper.instance().setDeepDifferentState();
-        expect(updateSpy).to.have.property('callCount', 1);
-      });
-
-      describeIf(is('>= 16.3'), 'setProps calls `componentDidUpdate` when `getDerivedStateFromProps` is defined', () => {
-        class DummyComp extends PureComponent {
-          constructor(...args) {
-            super(...args);
-            this.state = { state: -1 };
-          }
-
-          static getDerivedStateFromProps({ changeState, counter }) {
-            return changeState ? { state: counter * 10 } : null;
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { counter } = this.props;
-            const { state } = this.state;
-            return (
-              <p>
-                {counter}
-                {state}
-              </p>
-            );
-          }
-        }
-
-        let cDU;
-        let gDSFP;
-
-        beforeEach(() => { // eslint-disable-line mocha/no-sibling-hooks
-          cDU = sinon.spy(DummyComp.prototype, 'componentDidUpdate');
-          gDSFP = sinon.spy(DummyComp, 'getDerivedStateFromProps');
-        });
-
-        it('with no state changes, calls both methods with a sync and async setProps', () => {
-          const wrapper = mount(<DummyComp changeState={false} counter={0} />);
-
-          expect(cDU).to.have.property('callCount', 0);
-          expect(gDSFP).to.have.property('callCount', 1);
-          const [firstCall] = gDSFP.args;
-          expect(firstCall).to.eql([{
-            changeState: false,
-            counter: 0,
-          }, {
-            state: -1,
-          }]);
-          expect(wrapper.state()).to.eql({ state: -1 });
-
-          wrapper.setProps({ counter: 1 });
-
-          expect(cDU).to.have.property('callCount', 1);
-          expect(gDSFP).to.have.property('callCount', 2);
-          const [, secondCall] = gDSFP.args;
-          expect(secondCall).to.eql([{
-            changeState: false,
-            counter: 1,
-          }, {
-            state: -1,
-          }]);
-          expect(wrapper.state()).to.eql({ state: -1 });
-
-          return new Promise((resolve) => {
-            wrapper.setProps({ counter: 2 }, resolve);
-          }).then(() => {
-            expect(cDU).to.have.property('callCount', 2);
-            expect(gDSFP).to.have.property('callCount', 3);
-            const [, , thirdCall] = gDSFP.args;
-            expect(thirdCall).to.eql([{
-              changeState: false,
-              counter: 2,
-            }, {
-              state: -1,
-            }]);
-            expect(wrapper.state()).to.eql({ state: -1 });
-          });
-        });
-
-        it('with a state changes, calls both methods with a sync and async setProps', () => {
-          const wrapper = mount(<DummyComp changeState counter={0} />);
-
-          expect(gDSFP).to.have.property('callCount', 1);
-          const [firstCall] = gDSFP.args;
-          expect(firstCall).to.eql([{
-            changeState: true,
-            counter: 0,
-          }, {
-            state: -1,
-          }]);
-          expect(wrapper.state()).to.eql({ state: 0 });
-
-          wrapper.setProps({ counter: 1 });
-
-          expect(cDU).to.have.property('callCount', 1);
-          expect(gDSFP).to.have.property('callCount', 2);
-          const [, secondCall] = gDSFP.args;
-          expect(secondCall).to.eql([{
-            changeState: true,
-            counter: 1,
-          }, {
-            state: 0,
-          }]);
-          expect(wrapper.state()).to.eql({ state: 10 });
-
-          return new Promise((resolve) => {
-            wrapper.setProps({ counter: 2 }, resolve);
-          }).then(() => {
-            expect(cDU).to.have.property('callCount', 2);
-            expect(gDSFP).to.have.property('callCount', 3);
-            const [, , thirdCall] = gDSFP.args;
-            expect(thirdCall).to.eql([{
-              changeState: true,
-              counter: 2,
-            }, {
-              state: 10,
-            }]);
-            expect(wrapper.state()).to.eql({ state: 20 });
-          });
-        });
-      });
-
-      it('does not infinitely loop when a PureComponent fires a noop setState in cDU', () => {
-        class Example extends PureComponent {
-          constructor(props) {
-            super(props);
-
-            this.renders = 0;
-            this.state = {
-              a: false,
-              b: false,
-            };
-          }
-
-          componentDidMount() {
-            this.setState({ b: false });
-          }
-
-          componentDidUpdate() {
-            this.setState({ b: false }); // eslint-disable-line react/no-did-update-set-state
-          }
-
-          render() {
-            this.renders += 1;
-            const { a, b } = this.state;
-            return <div>{`${a} ${b} ${this.renders}`}</div>;
-          }
-        }
-
-        const wrapper = mount(<Example />);
-        expect(wrapper.debug()).to.equal(`<Example>
-  <div>
-    false false 1
-  </div>
-</Example>`);
-      });
-    });
-
-    describe('Own PureComponent implementation', () => {
-      it('does not update when state and props did not change', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          shouldComponentUpdate(nextProps, nextState) {
-            return !isEqual(this.props, nextProps) || !isEqual(this.state, nextState);
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-        const wrapper = mount(<Foo id={1} />);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-      });
-    });
-
-    describeIf(is('>= 16.3'), 'support getSnapshotBeforeUpdate', () => {
-      it('calls getSnapshotBeforeUpdate and pass snapshot to componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'bar',
-            };
-          }
-
-          componentDidUpdate(prevProps, prevState, snapshot) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, snapshot);
-          }
-
-          getSnapshotBeforeUpdate(prevProps, prevState) {
-            spy('getSnapshotBeforeUpdate', prevProps, this.props, prevState, this.state);
-            return { snapshot: 'ok' };
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-        const wrapper = mount(<Foo name="foo" />);
-        spy.resetHistory();
-        wrapper.setProps({ name: 'bar' });
-        expect(spy.args).to.deep.equal([
-          ['render'],
-          ['getSnapshotBeforeUpdate', { name: 'foo' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'bar' }],
-          ['componentDidUpdate', { name: 'foo' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'bar' }, { snapshot: 'ok' }],
-        ]);
-        spy.resetHistory();
-        wrapper.setState({ foo: 'baz' });
-        expect(spy.args).to.deep.equal([
-          ['render'],
-          ['getSnapshotBeforeUpdate', { name: 'bar' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'baz' }],
-          ['componentDidUpdate', { name: 'bar' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'baz' }, { snapshot: 'ok' }],
-        ]);
-      });
-    });
-  });
-
   it('works with class components that return null', () => {
     class Foo extends React.Component {
       render() {
@@ -3897,134 +1923,6 @@ describeWithDOM('mount', () => {
         onClick: divProps.onClick,
       });
       expect(divProps.onClick).to.be.a('function');
-    });
-  });
-
-  describe('lifecycles', () => {
-    it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-        }
-
-        componentDidUpdate() {}
-
-        onChange() {
-          this.setState({ foo: 'onChange update' });
-        }
-
-        render() {
-          const { foo } = this.state;
-          return <div>{foo}</div>;
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-      const wrapper = mount(<Foo />);
-      wrapper.setState({ foo: 'wrapper setState update' });
-      expect(wrapper.state('foo')).to.equal('wrapper setState update');
-      expect(spy).to.have.property('callCount', 1);
-      wrapper.instance().onChange();
-      expect(wrapper.state('foo')).to.equal('onChange update');
-      expect(spy).to.have.property('callCount', 2);
-    });
-
-    it('calls `componentDidUpdate` when component’s `setState` is called through a bound method', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-          this.onChange = this.onChange.bind(this);
-        }
-
-        componentDidUpdate() {}
-
-        onChange() {
-          // enzyme can't handle the update because `this` is a ReactComponent instance,
-          // not a Wrapper instance.
-          this.setState({ foo: 'onChange update' });
-        }
-
-        render() {
-          const { foo } = this.state;
-          return (
-            <div>
-              {foo}
-              <button type="button" onClick={this.onChange}>click</button>
-            </div>
-          );
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-      const wrapper = mount(<Foo />);
-      wrapper.find('button').prop('onClick')();
-      expect(wrapper.state('foo')).to.equal('onChange update');
-      expect(spy).to.have.property('callCount', 1);
-    });
-
-    it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-          this.update = () => this.setState({ foo: 'update' });
-        }
-
-        componentDidMount() {
-          this.update();
-        }
-
-        componentDidUpdate() {}
-
-        render() {
-          const { foo } = this.state;
-          return <div>{foo}</div>;
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-      const wrapper = mount(<Foo />);
-      expect(spy).to.have.property('callCount', 1);
-      expect(wrapper.state('foo')).to.equal('update');
-    });
-
-    it('does not call `componentDidMount` twice when a child component is created', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-        }
-
-        componentDidMount() {}
-
-        render() {
-          const { foo } = this.state;
-          return (
-            <div>
-              <button type="button" onClick={() => this.setState({ foo: 'update2' })}>
-                click
-              </button>
-              {foo}
-            </div>
-          );
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidMount');
-
-      const wrapper = mount(<Foo />);
-      expect(spy).to.have.property('callCount', 1);
-      wrapper.find('button').prop('onClick')();
-      expect(spy).to.have.property('callCount', 1);
     });
   });
 });
