+++ packages/enzyme-test-suite/test/ShallowWrapper-spec.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import { expect } from 'chai';
 import sinon from 'sinon-sandbox';
 import wrap from 'mocha-wrap';
-import isEqual from 'lodash.isequal';
 import inspect from 'object-inspect';
 
 import {
@@ -28,7 +27,6 @@ import {
   lazy,
   memo,
   Profiler,
-  PureComponent,
   Suspense,
 } from './_helpers/react-compat';
 import {
@@ -36,15 +34,11 @@ import {
   itIf,
 } from './_helpers';
 import describeMethods from './_helpers/describeMethods';
+import describeLifecycles from './_helpers/describeLifecycles';
 import describeHooks from './_helpers/describeHooks';
 import {
-  REACT16,
   is,
 } from './_helpers/version';
-
-// The shallow renderer in react 16 does not yet support batched updates. When it does,
-// we should be able to go un-skip all of the tests that are skipped with this flag.
-const BATCHING = !REACT16;
 
 describe('shallow', () => {
   describe('top level entry points', () => {
@@ -1326,6 +1320,17 @@ describe('shallow', () => {
     'unmount',
     'wrap',
   );
+  describeLifecycles(
+    { Wrap, Wrapper },
+    'componentDidCatch',
+    'componentDidMount',
+    'componentDidUpdate',
+    'componentWillUnmount',
+    'getDerivedStateFromError',
+    'getDerivedStateFromProps',
+    'getSnapshotBeforeUpdate',
+    'misc',
+  );
   describeHooks(
     { Wrap, Wrapper },
     'useCallback',
@@ -2333,1913 +2338,6 @@ describe('shallow', () => {
         );
       });
     });
-
-    describeIf(is('>= 16.3'), 'getDerivedStateFromProps', () => {
-      let spy;
-
-      beforeEach(() => {
-        spy = sinon.spy();
-      });
-
-      class Spy extends React.Component {
-        constructor(...args) {
-          super(...args);
-          this.state = { state: true }; // eslint-disable-line react/no-unused-state
-          spy('constructor');
-        }
-
-        shouldComponentUpdate(nextProps, nextState, nextContext) {
-          spy('shouldComponentUpdate', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState,
-            prevContext: this.context,
-            nextContext,
-          });
-          return true;
-        }
-
-        componentWillUpdate(nextProps, nextState, nextContext) {
-          spy('componentWillUpdate', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState,
-            prevContext: this.context,
-            nextContext,
-          });
-        }
-
-        componentDidUpdate(prevProps, prevState, prevContext) {
-          spy('componentDidUpdate', {
-            prevProps,
-            nextProps: this.props,
-            prevState,
-            nextState: this.state,
-            prevContext,
-            nextContext: this.context,
-          });
-        }
-
-        render() {
-          spy('render');
-          return null;
-        }
-      }
-
-      class CWRP extends Spy {
-        componentWillReceiveProps(nextProps, nextContext) {
-          spy('componentWillReceiveProps', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState: this.state,
-            prevContext: this.context,
-            nextContext,
-          });
-        }
-      }
-
-      class U_CWRP extends Spy {
-        UNSAFE_componentWillReceiveProps(nextProps) { // eslint-disable-line camelcase
-          spy('UNSAFE_componentWillReceiveProps', {
-            prevProps: this.props,
-            nextProps,
-            prevState: this.state,
-            nextState: this.state,
-            prevContext: this.context,
-            nextContext: undefined,
-          });
-        }
-      }
-
-      class GDSFP extends Spy {
-        static getDerivedStateFromProps(props, state) {
-          spy('getDerivedStateFromProps', { props, state });
-          return {};
-        }
-      }
-
-      it('calls cWRP when expected', () => {
-        const prevProps = { a: 1 };
-        const wrapper = shallow(<CWRP {...prevProps} />);
-        expect(spy.args).to.deep.equal([
-          ['constructor'],
-          ['render'],
-        ]);
-        spy.resetHistory();
-
-        const foo = {};
-        const props = { foo };
-        const {
-          context: prevContext,
-          context: nextContext,
-          state: prevState,
-          state: nextState,
-        } = wrapper.instance();
-
-        wrapper.setProps(props);
-        const nextProps = { ...prevProps, ...props };
-
-        const data = {
-          prevProps,
-          nextProps,
-          prevState,
-          nextState,
-          prevContext,
-          nextContext,
-        };
-        expect(spy.args).to.deep.equal([
-          ['componentWillReceiveProps', data],
-          ['shouldComponentUpdate', data],
-          ['componentWillUpdate', data],
-          ['render'],
-          ['componentDidUpdate', {
-            ...data,
-            prevContext: is('>= 16') ? undefined : prevContext,
-          }],
-        ]);
-      });
-
-      it('calls UNSAFE_cWRP when expected', () => {
-        const prevProps = { a: 1 };
-        // eslint-disable-next-line react/jsx-pascal-case
-        const wrapper = shallow(<U_CWRP {...prevProps} />);
-        expect(spy.args).to.deep.equal([
-          ['constructor'],
-          ['render'],
-        ]);
-        spy.resetHistory();
-
-        const foo = {};
-        const props = { foo };
-        const {
-          context: prevContext,
-          context: nextContext,
-          state: prevState,
-          state: nextState,
-        } = wrapper.instance();
-
-        wrapper.setProps(props);
-        const nextProps = { ...prevProps, ...props };
-
-        const data = {
-          prevProps,
-          nextProps,
-          prevState,
-          nextState,
-          prevContext,
-          nextContext,
-        };
-        expect(spy.args).to.deep.equal([
-          ['UNSAFE_componentWillReceiveProps', {
-            ...data,
-            nextContext: is('>= 16') ? undefined : nextContext,
-          }],
-          ['shouldComponentUpdate', data],
-          ['componentWillUpdate', data],
-          ['render'],
-          ['componentDidUpdate', {
-            ...data,
-            prevContext: is('>= 16') ? undefined : prevContext,
-          }],
-        ]);
-      });
-
-      it('calls gDSFP when expected', () => {
-        const prevProps = { a: 1 };
-        const state = { state: true };
-        const wrapper = shallow(<GDSFP {...prevProps} />);
-        expect(spy.args).to.deep.equal([
-          ['constructor'],
-          ['getDerivedStateFromProps', {
-            props: prevProps,
-            state,
-          }],
-          ['render'],
-        ]);
-        spy.resetHistory();
-
-        const foo = {};
-        const props = { foo };
-        const {
-          context: prevContext,
-          context: nextContext,
-          state: prevState,
-          state: nextState,
-        } = wrapper.instance();
-
-        wrapper.setProps(props);
-        const nextProps = { ...prevProps, ...props };
-
-        const data = {
-          prevProps,
-          nextProps,
-          prevState,
-          nextState,
-          prevContext,
-          nextContext,
-        };
-        expect(spy.args).to.deep.equal([
-          ['getDerivedStateFromProps', {
-            props: nextProps,
-            state: nextState,
-          }],
-          ['shouldComponentUpdate', data],
-          ['render'],
-          ['componentDidUpdate', {
-            ...data,
-            prevContext: is('>= 16') ? undefined : prevContext,
-          }],
-        ]);
-      });
-
-      it('cDU’s nextState differs from `this.state` when gDSFP returns new state', () => {
-        class SimpleComponent extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = { value: props.value };
-          }
-
-          static getDerivedStateFromProps(props, state) {
-            return props.value === state.value ? null : { value: props.value };
-          }
-
-          shouldComponentUpdate(nextProps, nextState) {
-            const { value } = this.state;
-            return nextState.value !== value;
-          }
-
-          render() {
-            const { value } = this.state;
-            return (<input value={value} />);
-          }
-        }
-        const wrapper = shallow(<SimpleComponent value="initial" />);
-
-        expect(wrapper.find('input').prop('value')).to.equal('initial');
-
-        wrapper.setProps({ value: 'updated' });
-
-        expect(wrapper.find('input').prop('value')).to.equal('updated');
-      });
-    });
-
-    describeIf(is('>= 16'), 'componentDidCatch', () => {
-      describe('errors inside an error boundary', () => {
-        const errorToThrow = new EvalError('threw an error!');
-
-        const hasFragments = is('>= 16.2');
-        const MaybeFragment = hasFragments ? Fragment : 'main';
-
-        function Thrower({ throws }) {
-          if (throws) {
-            throw errorToThrow;
-          }
-          return null;
-        }
-
-        class ErrorBoundary extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              throws: false,
-              didThrow: false,
-            };
-          }
-
-          componentDidCatch(error, info) {
-            const { spy } = this.props;
-            spy(error, info);
-            this.setState({
-              throws: false,
-              didThrow: true,
-            });
-          }
-
-          render() {
-            const {
-              didThrow,
-              throws,
-            } = this.state;
-            return (
-              <div>
-                <MaybeFragment>
-                  <span>
-                    <Thrower throws={throws} />
-                    <div>
-                      {didThrow ? 'HasThrown' : 'HasNotThrown'}
-                    </div>
-                  </span>
-                </MaybeFragment>
-              </div>
-            );
-          }
-        }
-
-        describe('Thrower', () => {
-          it('does not throw when `throws` is `false`', () => {
-            expect(() => shallow(<Thrower throws={false} />)).not.to.throw();
-          });
-
-          it('throws when `throws` is `true`', () => {
-            expect(() => shallow(<Thrower throws />)).to.throw(errorToThrow);
-          });
-        });
-
-        it('catches a simulated error', () => {
-          const spy = sinon.spy();
-          const wrapper = shallow(<ErrorBoundary spy={spy} />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError, info]] = spy.args;
-          expect(() => { throw actualError; }).to.throw(errorToThrow);
-          expect(info).to.deep.equal({
-            componentStack: `
-    in Thrower (created by ErrorBoundary)
-    in span (created by ErrorBoundary)${hasFragments ? '' : `
-    in main (created by ErrorBoundary)`}
-    in div (created by ErrorBoundary)
-    in ErrorBoundary (created by WrapperComponent)
-    in WrapperComponent`,
-          });
-        });
-
-        it('rerenders on a simulated error', () => {
-          const wrapper = shallow(<ErrorBoundary spy={sinon.stub()} />);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(1);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(0);
-        });
-
-        it('does not catch errors during shallow render', () => {
-          const spy = sinon.spy();
-          const wrapper = shallow(<ErrorBoundary spy={spy} />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          wrapper.setState({ throws: true });
-
-          expect(spy).to.have.property('callCount', 0);
-
-          const thrower = wrapper.find(Thrower);
-          expect(thrower).to.have.lengthOf(1);
-          expect(thrower.props()).to.have.property('throws', true);
-
-          expect(() => thrower.dive()).to.throw(errorToThrow);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-        });
-      });
-    });
-
-    describeIf(is('>= 16.6'), 'getDerivedStateFromError', () => {
-      describe('errors inside an error boundary', () => {
-        const errorToThrow = new EvalError('threw an error!');
-
-        function Thrower({ throws }) {
-          if (throws) {
-            throw errorToThrow;
-          }
-          return null;
-        }
-
-        function getErrorBoundary() {
-          return class ErrorBoundary extends React.Component {
-            static getDerivedStateFromError() {
-              return {
-                throws: false,
-                didThrow: true,
-              };
-            }
-
-            constructor(props) {
-              super(props);
-              this.state = {
-                throws: false,
-                didThrow: false,
-              };
-            }
-
-            render() {
-              const {
-                didThrow,
-                throws,
-              } = this.state;
-
-              return (
-                <div>
-                  <Fragment>
-                    <span>
-                      <Thrower throws={throws} />
-                      <div>
-                        {didThrow ? 'HasThrown' : 'HasNotThrown'}
-                      </div>
-                    </span>
-                  </Fragment>
-                </div>
-              );
-            }
-          };
-        }
-
-        describe('Thrower', () => {
-          it('does not throw when `throws` is `false`', () => {
-            expect(() => shallow(<Thrower throws={false} />)).not.to.throw();
-          });
-
-          it('throws when `throws` is `true`', () => {
-            expect(() => shallow(<Thrower throws />)).to.throw();
-          });
-        });
-
-        it('catches a simulated error', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const spy = sinon.spy(ErrorBoundary, 'getDerivedStateFromError');
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(spy).to.have.property('callCount', 1);
-
-          expect(spy.args).to.be.an('array').and.have.lengthOf(1);
-          const [[actualError]] = spy.args;
-          expect(actualError).to.equal(errorToThrow);
-        });
-
-        it('rerenders on a simulated error', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(1);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(0);
-        });
-
-        it('does not catch errors during shallow render', () => {
-          const ErrorBoundary = getErrorBoundary();
-
-          const spy = sinon.spy(ErrorBoundary, 'getDerivedStateFromError');
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          wrapper.setState({ throws: true });
-
-          expect(spy).to.have.property('callCount', 0);
-
-          const thrower = wrapper.find(Thrower);
-          expect(thrower).to.have.lengthOf(1);
-          expect(thrower.props()).to.have.property('throws', true);
-
-          expect(() => thrower.dive()).to.throw(errorToThrow);
-
-          expect(spy).to.have.property('callCount', 0);
-
-          expect(wrapper.find({ children: 'HasThrown' })).to.have.lengthOf(0);
-          expect(wrapper.find({ children: 'HasNotThrown' })).to.have.lengthOf(1);
-        });
-      });
-    });
-
-    describeIf(is('>= 16.6'), 'getDerivedStateFromError and componentDidCatch combined', () => {
-
-      const errorToThrow = new EvalError('threw an error!');
-      const expectedInfo = {
-        componentStack: `
-    in Thrower (created by ErrorBoundary)
-    in div (created by ErrorBoundary)
-    in ErrorBoundary (created by WrapperComponent)
-    in WrapperComponent`,
-      };
-
-      function Thrower({ throws }) {
-        if (throws) {
-          throw errorToThrow;
-        }
-        return null;
-      }
-
-      describe('errors inside error boundary when getDerivedStateFromProps returns update', () => {
-        let lifecycleSpy;
-        let stateSpy;
-
-        beforeEach(() => {
-          lifecycleSpy = sinon.spy();
-          stateSpy = sinon.spy();
-        });
-
-        class ErrorBoundary extends React.Component {
-          static getDerivedStateFromError(error) {
-            lifecycleSpy('getDerivedStateFromError', error);
-            return {
-              didThrow: true,
-              throws: false,
-            };
-          }
-
-          constructor(props) {
-            super(props);
-            this.state = {
-              didThrow: false,
-              throws: false,
-            };
-
-            lifecycleSpy('constructor');
-          }
-
-          componentDidCatch(error, info) {
-            lifecycleSpy('componentDidCatch', error, info);
-            stateSpy({ ...this.state });
-          }
-
-          render() {
-            lifecycleSpy('render');
-
-            const {
-              throws,
-            } = this.state;
-
-            return (
-              <div>
-                <Thrower throws={throws} />
-              </div>
-            );
-          }
-        }
-
-        it('does not catch errors during shallow render', () => {
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(lifecycleSpy).to.have.property('callCount', 2);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          expect(stateSpy).to.have.property('callCount', 0);
-
-          lifecycleSpy.resetHistory();
-
-          wrapper.setState({ throws: true });
-
-          const thrower = wrapper.find(Thrower);
-          expect(thrower).to.have.lengthOf(1);
-          expect(thrower.props()).to.have.property('throws', true);
-
-          expect(() => thrower.dive()).to.throw(errorToThrow);
-
-          expect(lifecycleSpy).to.have.property('callCount', 1);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['render'],
-          ]);
-        });
-
-        it('calls getDerivedStateFromError first and then componentDidCatch for simulated error', () => {
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(lifecycleSpy).to.have.property('callCount', 2);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          expect(stateSpy).to.have.property('callCount', 0);
-
-          lifecycleSpy.resetHistory();
-
-          expect(() => wrapper.find(Thrower).simulateError(errorToThrow)).not.to.throw();
-
-          expect(lifecycleSpy).to.have.property('callCount', 3);
-          expect(lifecycleSpy.args).to.deep.equal([
-            ['getDerivedStateFromError', errorToThrow],
-            ['render'],
-            ['componentDidCatch', errorToThrow, expectedInfo],
-          ]);
-
-          expect(stateSpy).to.have.property('callCount', 1);
-          expect(stateSpy.args).to.deep.equal([
-            [{
-              throws: false,
-              didThrow: true,
-            }],
-          ]);
-        });
-      });
-
-      describe('errors inside error boundary when getDerivedStateFromError does not return update', () => {
-        let spy;
-
-        beforeEach(() => {
-          spy = sinon.spy();
-        });
-
-        class ErrorBoundary extends React.Component {
-          static getDerivedStateFromError(error) {
-            spy('getDerivedStateFromError', error);
-            return null;
-          }
-
-          constructor(props) {
-            super(props);
-            this.state = {
-              didThrow: false,
-              throws: false,
-            };
-
-            spy('constructor');
-          }
-
-          componentDidCatch(error, info) {
-            spy('componentDidCatch', error, info);
-
-            this.setState({
-              didThrow: true,
-              throws: false,
-            });
-          }
-
-          render() {
-            spy('render');
-
-            const {
-              didThrow,
-              throws,
-            } = this.state;
-
-            return (
-              <div>
-                <Thrower throws={throws} />
-                <div>
-                  {didThrow ? 'HasThrown' : 'HasNotThrown'}
-                </div>
-              </div>
-            );
-          }
-        }
-
-        it('does not catch errors during shallow render', () => {
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 2);
-          expect(spy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          spy.resetHistory();
-
-          wrapper.setState({ throws: true });
-
-          const thrower = wrapper.find(Thrower);
-          expect(thrower).to.have.lengthOf(1);
-          expect(thrower.props()).to.have.property('throws', true);
-
-          expect(() => thrower.dive()).to.throw(errorToThrow);
-
-          expect(spy).to.have.property('callCount', 1);
-          expect(spy.args).to.deep.equal([
-            ['render'],
-          ]);
-        });
-
-        it('rerenders on a simulated error', () => {
-          const wrapper = shallow(<ErrorBoundary />);
-
-          expect(spy).to.have.property('callCount', 2);
-          expect(spy.args).to.deep.equal([
-            ['constructor'],
-            ['render'],
-          ]);
-
-          spy.resetHistory();
-
-          const thrower = wrapper.find(Thrower);
-
-          expect(() => thrower.simulateError(errorToThrow)).not.to.throw(errorToThrow);
-
-          expect(spy).to.have.property('callCount', 3);
-          expect(spy.args).to.deep.equal([
-            ['getDerivedStateFromError', errorToThrow],
-            ['componentDidCatch', errorToThrow, expectedInfo],
-            ['render'],
-          ]);
-        });
-      });
-    });
-
-    context('mounting phase', () => {
-      it('calls componentWillMount and componentDidMount', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          componentWillMount() {
-            spy('componentWillMount');
-          }
-
-          componentDidMount() {
-            spy('componentDidMount');
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-        shallow(<Foo />);
-        expect(spy.args).to.deep.equal([
-          ['componentWillMount'],
-          ['render'],
-          ['componentDidMount'],
-        ]);
-      });
-
-      itIf(BATCHING, 'is batching updates', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillMount() {
-            /* eslint-disable react/destructuring-assignment */
-            this.setState({ count: this.state.count + 1 });
-            this.setState({ count: this.state.count + 1 });
-            /* eslint-enable react/destructuring-assignment */
-          }
-
-          componentDidMount() {
-            /* eslint-disable react/destructuring-assignment */
-            this.setState({ count: this.state.count + 1 });
-            this.setState({ count: this.state.count + 1 });
-            /* eslint-enable react/destructuring-assignment */
-          }
-
-          render() {
-            spy();
-            const { count } = this.state;
-            return <div>{count}</div>;
-          }
-        }
-        const result = shallow(<Foo />);
-        expect(result.state('count')).to.equal(2);
-        expect(spy).to.have.property('callCount', 2);
-      });
-    });
-
-    context('updating props', () => {
-      it('calls shouldComponentUpdate, componentWillUpdate, and componentDidUpdate', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              foo: 'state',
-            };
-          }
-
-          componentWillReceiveProps(nextProps, nextContext) {
-            spy('componentWillReceiveProps', this.props, nextProps, nextContext);
-          }
-
-          shouldComponentUpdate(nextProps, nextState, nextContext) {
-            spy('shouldComponentUpdate', this.props, nextProps, this.state, nextState, nextContext);
-            return true;
-          }
-
-          componentWillUpdate(nextProps, nextState, nextContext) {
-            spy('componentWillUpdate', this.props, nextProps, this.state, nextState, nextContext);
-          }
-
-          componentDidUpdate(prevProps, prevState, prevContext) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, prevContext);
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-
-        const wrapper = shallow(
-          <Foo foo="bar" />,
-          {
-            context: { foo: 'context' },
-          },
-        );
-        wrapper.setProps({ foo: 'baz' });
-        wrapper.setProps({ foo: 'bax' });
-        expect(spy.args).to.deep.equal([
-          [
-            'render',
-          ],
-          [
-            'componentWillReceiveProps',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'context' }, // this will be fixed
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'state' }, { foo: 'state' },
-            is('>= 16') ? undefined : { foo: 'context' },
-          ],
-          [
-            'componentWillReceiveProps',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'context' },
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'context' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'baz' }, { foo: 'bax' },
-            { foo: 'state' }, { foo: 'state' },
-            is('>= 16') ? undefined : { foo: 'context' },
-          ],
-        ]);
-      });
-
-      it('calls componentWillReceiveProps, shouldComponentUpdate, componentWillUpdate and componentDidUpdate with merged props', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          componentWillReceiveProps(nextProps) {
-            spy('componentWillReceiveProps', this.props, nextProps);
-          }
-
-          shouldComponentUpdate(nextProps) {
-            spy('shouldComponentUpdate', this.props, nextProps);
-            return true;
-          }
-
-          componentWillUpdate(nextProps) {
-            spy('componentWillUpdate', this.props, nextProps);
-          }
-
-          componentDidUpdate(prevProps) {
-            spy('componentDidUpdate', prevProps, this.props);
-          }
-
-          render() {
-            return (
-              <div />
-            );
-          }
-        }
-
-        const wrapper = shallow(<Foo a="a" b="b" />);
-
-        wrapper.setProps({ b: 'c', d: 'e' });
-
-        expect(spy.args).to.deep.equal([
-          [
-            'componentWillReceiveProps',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-          [
-            'shouldComponentUpdate',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-          [
-            'componentWillUpdate',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-          [
-            'componentDidUpdate',
-            { a: 'a', b: 'b' },
-            { a: 'a', b: 'c', d: 'e' },
-          ],
-        ]);
-      });
-
-      it('cancels rendering when Component returns false in shouldComponentUpdate', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          componentWillReceiveProps() {
-            spy('componentWillReceiveProps');
-          }
-
-          shouldComponentUpdate() {
-            spy('shouldComponentUpdate');
-            return false;
-          }
-
-          componentWillUpdate() {
-            spy('componentWillUpdate');
-          }
-
-          componentDidUpdate() {
-            spy('componentDidUpdate');
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-
-        const wrapper = shallow(<Foo foo="bar" />);
-        expect(wrapper.instance().props.foo).to.equal('bar');
-        expect(spy.args).to.deep.equal([
-          ['render'],
-        ]);
-
-        spy.resetHistory();
-        wrapper.setProps({ foo: 'baz' });
-        expect(wrapper.instance().props.foo).to.equal('baz');
-        expect(spy.args).to.deep.equal([
-          ['componentWillReceiveProps'],
-          ['shouldComponentUpdate'],
-        ]);
-
-        spy.resetHistory();
-        wrapper.setProps({ foo: 'bax' });
-        expect(wrapper.instance().props.foo).to.equal('bax');
-        expect(spy.args).to.deep.equal([
-          ['componentWillReceiveProps'],
-          ['shouldComponentUpdate'],
-        ]);
-      });
-
-      itIf(BATCHING, 'does not provoke another renders to call setState in componentWillReceiveProps', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillReceiveProps() {
-            /* eslint-disable react/destructuring-assignment */
-            this.setState({ count: this.state.count + 1 });
-            this.setState({ count: this.state.count + 1 });
-            /* eslint-enable react/destructuring-assignment */
-          }
-
-          render() {
-            spy();
-            const { foo } = this.props;
-            return <div>{foo}</div>;
-          }
-        }
-        const result = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setProps({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 2);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      itIf(BATCHING, 'provokes an another render to call setState twice in componentWillUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { foo } = this.props;
-            return <div>{foo}</div>;
-          }
-        }
-        const result = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setProps({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      itIf(BATCHING, 'provokes an another render to call setState twice in componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentDidUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/no-did-update-set-state, react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/no-did-update-set-state, react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { foo } = this.props;
-            return <div>{foo}</div>;
-          }
-        }
-        const result = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setProps({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-    });
-
-    context('updating state', () => {
-      it('calls shouldComponentUpdate, componentWillUpdate and componentDidUpdate', () => {
-        const spy = sinon.spy();
-
-        class Foo extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              foo: 'bar',
-            };
-          }
-
-          shouldComponentUpdate(nextProps, nextState, nextContext) {
-            spy('shouldComponentUpdate', this.props, nextProps, this.state, nextState, nextContext);
-            return true;
-          }
-
-          componentWillUpdate(nextProps, nextState, nextContext) {
-            spy('componentWillUpdate', this.props, nextProps, this.state, nextState, nextContext);
-          }
-
-          componentDidUpdate(prevProps, prevState, prevContext) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, prevContext);
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-
-        const wrapper = shallow(
-          <Foo foo="props" />,
-          {
-            context: { foo: 'context' },
-          },
-        );
-        wrapper.setState({ foo: 'baz' });
-        const expected = [
-          [
-            'render',
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'context' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'bar' }, { foo: 'baz' },
-            { foo: 'context' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'bar' }, { foo: 'baz' },
-            is('>= 16') ? undefined : { foo: 'context' },
-          ],
-        ];
-        expect(spy.args).to.deep.equal(expected);
-      });
-
-      it('cancels rendering when Component returns false in shouldComponentUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'bar',
-            };
-          }
-
-          shouldComponentUpdate() {
-            spy('shouldComponentUpdate');
-            return false;
-          }
-
-          componentWillUpdate() {
-            spy('componentWillUpdate');
-          }
-
-          componentDidUpdate() {
-            spy('componentDidUpdate');
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const wrapper = shallow(<Foo />);
-        expect(wrapper.instance().state.foo).to.equal('bar');
-        wrapper.setState({ foo: 'baz' });
-        expect(wrapper.instance().state.foo).to.equal('baz');
-        expect(spy.args).to.deep.equal([['render'], ['shouldComponentUpdate']]);
-      });
-
-      itIf(BATCHING, 'provokes an another render to call setState twice in componentWillUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              name: 'foo',
-              count: 0,
-            };
-          }
-
-          componentWillUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setState({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      itIf(BATCHING, 'provokes an another render to call setState twice in componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              name: 'foo',
-              count: 0,
-            };
-          }
-
-          componentDidUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/no-did-update-set-state, react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/no-did-update-set-state, react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        result.setState({ name: 'bar' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-    });
-
-    context('updating context', () => {
-      it('calls shouldComponentUpdate, componentWillUpdate and componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(...args) {
-            super(...args);
-            this.state = {
-              foo: 'state',
-            };
-          }
-
-          shouldComponentUpdate(nextProps, nextState, nextContext) {
-            spy('shouldComponentUpdate', this.props, nextProps, this.state, nextState, nextContext);
-            return true;
-          }
-
-          componentWillUpdate(nextProps, nextState, nextContext) {
-            spy('componentWillUpdate', this.props, nextProps, this.state, nextState, nextContext);
-          }
-
-          componentDidUpdate(prevProps, prevState, prevContext) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, prevContext);
-          }
-
-          render() {
-            spy('render');
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-        const wrapper = shallow(
-          <Foo foo="props" />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        expect(wrapper.instance().context.foo).to.equal('bar');
-        wrapper.setContext({ foo: 'baz' });
-        expect(wrapper.instance().context.foo).to.equal('baz');
-        expect(spy.args).to.deep.equal([
-          [
-            'render',
-          ],
-          [
-            'shouldComponentUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'baz' },
-          ],
-          [
-            'componentWillUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'state' }, { foo: 'state' },
-            { foo: 'baz' },
-          ],
-          [
-            'render',
-          ],
-          [
-            'componentDidUpdate',
-            { foo: 'props' }, { foo: 'props' },
-            { foo: 'state' }, { foo: 'state' },
-            is('>= 16') ? undefined : { foo: 'bar' },
-          ],
-        ]);
-      });
-
-      it('cancels rendering when Component returns false in shouldComponentUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          shouldComponentUpdate() {
-            spy('shouldComponentUpdate');
-            return false;
-          }
-
-          componentWillUpdate() {
-            spy('componentWillUpdate');
-          }
-
-          componentDidUpdate() {
-            spy('componentDidUpdate');
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-        Foo.contextTypes = {
-          foo: PropTypes.string,
-        };
-        const wrapper = shallow(
-          <Foo />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        wrapper.setContext({ foo: 'baz' });
-        expect(spy.args).to.deep.equal([['render'], ['shouldComponentUpdate']]);
-      });
-
-      itIf(BATCHING, 'provokes an another render to call setState twice in componentWillUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentWillUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = shallow(
-          <Foo />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        expect(spy).to.have.property('callCount', 1);
-        result.setContext({ foo: 'baz' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-
-      itIf(BATCHING, 'provokes an another render to call setState twice in componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.updated = false;
-            this.state = {
-              count: 0,
-            };
-          }
-
-          componentDidUpdate() {
-            if (!this.updated) {
-              this.updated = true;
-              /* eslint-disable react/no-did-update-set-state, react/destructuring-assignment */
-              this.setState({ count: this.state.count + 1 });
-              this.setState({ count: this.state.count + 1 });
-              /* eslint-enable react/no-did-update-set-state, react/destructuring-assignment */
-            }
-          }
-
-          render() {
-            spy();
-            const { name } = this.state;
-            return <div>{name}</div>;
-          }
-        }
-        const result = shallow(
-          <Foo />,
-          {
-            context: { foo: 'bar' },
-          },
-        );
-        expect(spy).to.have.property('callCount', 1);
-        result.setContext({ foo: 'baz' });
-        expect(spy).to.have.property('callCount', 3);
-        expect(result.state('count')).to.equal(1);
-      });
-    });
-
-    context('unmounting phase', () => {
-      it('calls componentWillUnmount', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          componentWillUnmount() {
-            spy();
-          }
-
-          render() {
-            return <div>foo</div>;
-          }
-        }
-        const wrapper = shallow(<Foo />);
-        wrapper.unmount();
-        expect(spy).to.have.property('callCount', 1);
-      });
-    });
-
-    context('component instance', () => {
-      it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          componentDidUpdate() {}
-
-          onChange() {
-            // enzyme can't handle the update because `this` is a ReactComponent instance,
-            // not a ShallowWrapper instance.
-            this.setState({ foo: 'onChange update' });
-          }
-
-          render() {
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-        const wrapper = shallow(<Foo />);
-        wrapper.setState({ foo: 'wrapper setState update' });
-        expect(wrapper.state('foo')).to.equal('wrapper setState update');
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.instance().onChange();
-        expect(wrapper.state('foo')).to.equal('onChange update');
-        expect(spy).to.have.property('callCount', 2);
-      });
-
-      it('calls `componentDidUpdate` when component’s `setState` is called through a bound method', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-            this.onChange = this.onChange.bind(this);
-          }
-
-          componentDidUpdate() {}
-
-          onChange() {
-            // enzyme can't handle the update because `this` is a ReactComponent instance,
-            // not a ShallowWrapper instance.
-            this.setState({ foo: 'onChange update' });
-          }
-
-          render() {
-            const { foo } = this.state;
-            return (
-              <div>
-                {foo}
-                <button type="button" onClick={this.onChange}>click</button>
-              </div>
-            );
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-        const wrapper = shallow(<Foo />);
-        wrapper.find('button').prop('onClick')();
-        expect(wrapper.state('foo')).to.equal('onChange update');
-        expect(spy).to.have.property('callCount', 1);
-      });
-
-      it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-            this.update = () => this.setState({ foo: 'update' });
-          }
-
-          componentDidMount() {
-            this.update();
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { foo } = this.state;
-            return <div>{foo}</div>;
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-        const wrapper = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        expect(wrapper.state('foo')).to.equal('update');
-      });
-
-      it('does not call `componentDidMount` twice when a child component is created', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          componentDidMount() {}
-
-          render() {
-            const { foo } = this.state;
-            return (
-              <div>
-                <button type="button" onClick={() => this.setState({ foo: 'update2' })}>
-                  click
-                </button>
-                {foo}
-              </div>
-            );
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidMount');
-
-        const wrapper = shallow(<Foo />);
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.find('button').prop('onClick')();
-        expect(spy).to.have.property('callCount', 1);
-      });
-    });
-
-    describeIf(is('>= 15.3'), 'PureComponent', () => {
-      it('does not update when state and props did not change', () => {
-        class Foo extends PureComponent {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { foo } = this.state;
-            return (
-              <div>
-                {foo}
-              </div>
-            );
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-        const wrapper = shallow(<Foo id={1} />);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-      });
-
-      class Test extends PureComponent {
-        constructor(...args) {
-          super(...args);
-
-          this.state = { a: { b: { c: 1 } } };
-        }
-
-        componentDidUpdate() {
-          const { onUpdate } = this.props;
-          onUpdate();
-        }
-
-        setDeepEqualState() {
-          this.setState({ a: { b: { c: 1 } } });
-        }
-
-        setDeepDifferentState() {
-          this.setState({ a: { b: { c: 2 } } });
-        }
-
-        render() {
-          const { a: { b: { c } } } = this.state;
-          return <div>{c}</div>;
-        }
-      }
-
-      it('rerenders on setState when new state is !==, but deeply equal to existing state', () => {
-        const updateSpy = sinon.spy();
-        const wrapper = shallow(<Test onUpdate={updateSpy} />);
-        wrapper.instance().setDeepEqualState();
-        expect(updateSpy).to.have.property('callCount', 1);
-      });
-
-      it('rerenders when setState is called with an object that doesnt have deep equality', () => {
-        const updateSpy = sinon.spy();
-        const wrapper = shallow(<Test onUpdate={updateSpy} />);
-        wrapper.instance().setDeepDifferentState();
-        expect(updateSpy).to.have.property('callCount', 1);
-      });
-
-      describeIf(is('>= 16.3'), 'setProps calls `componentDidUpdate` when `getDerivedStateFromProps` is defined', () => {
-        class DummyComp extends PureComponent {
-          constructor(...args) {
-            super(...args);
-            this.state = { state: -1 };
-          }
-
-          static getDerivedStateFromProps({ changeState, counter }) {
-            return changeState ? { state: counter * 10 } : null;
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { counter } = this.props;
-            const { state } = this.state;
-            return (
-              <p>
-                {counter}
-                {state}
-              </p>
-            );
-          }
-        }
-
-        let cDU;
-        let gDSFP;
-
-        beforeEach(() => { // eslint-disable-line mocha/no-sibling-hooks
-          cDU = sinon.spy(DummyComp.prototype, 'componentDidUpdate');
-          gDSFP = sinon.spy(DummyComp, 'getDerivedStateFromProps');
-        });
-
-        it('with no state changes, calls both methods with a sync and async setProps', () => {
-          const wrapper = shallow(<DummyComp changeState={false} counter={0} />);
-
-          expect(gDSFP).to.have.property('callCount', 1);
-          const [firstCall] = gDSFP.args;
-          expect(firstCall).to.eql([{
-            changeState: false,
-            counter: 0,
-          }, {
-            state: -1,
-          }]);
-          expect(wrapper.state()).to.eql({ state: -1 });
-
-          wrapper.setProps({ counter: 1 });
-
-          expect(cDU).to.have.property('callCount', 1);
-          expect(gDSFP).to.have.property('callCount', 2);
-          const [, secondCall] = gDSFP.args;
-          expect(secondCall).to.eql([{
-            changeState: false,
-            counter: 1,
-          }, {
-            state: -1,
-          }]);
-          expect(wrapper.state()).to.eql({ state: -1 });
-
-          return new Promise((resolve) => {
-            wrapper.setProps({ counter: 2 }, resolve);
-          }).then(() => {
-            expect(cDU).to.have.property('callCount', 2);
-            expect(gDSFP).to.have.property('callCount', 3);
-            const [, , thirdCall] = gDSFP.args;
-            expect(thirdCall).to.eql([{
-              changeState: false,
-              counter: 2,
-            }, {
-              state: -1,
-            }]);
-            expect(wrapper.state()).to.eql({ state: -1 });
-          });
-        });
-
-        it('with a state changes, calls both methods with a sync and async setProps', () => {
-          const wrapper = shallow(<DummyComp changeState counter={0} />);
-
-          expect(cDU).to.have.property('callCount', 0);
-          expect(gDSFP).to.have.property('callCount', 1);
-          const [firstCall] = gDSFP.args;
-          expect(firstCall).to.eql([{
-            changeState: true,
-            counter: 0,
-          }, {
-            state: -1,
-          }]);
-          expect(wrapper.state()).to.eql({ state: 0 });
-
-          wrapper.setProps({ counter: 1 });
-
-          expect(cDU).to.have.property('callCount', 1);
-          expect(gDSFP).to.have.property('callCount', 2);
-          const [, secondCall] = gDSFP.args;
-          expect(secondCall).to.eql([{
-            changeState: true,
-            counter: 1,
-          }, {
-            state: 0,
-          }]);
-          expect(wrapper.state()).to.eql({ state: 10 });
-
-          return new Promise((resolve) => {
-            wrapper.setProps({ counter: 2 }, resolve);
-          }).then(() => {
-            expect(cDU).to.have.property('callCount', 2);
-            expect(gDSFP).to.have.property('callCount', 3);
-            const [, , thirdCall] = gDSFP.args;
-            expect(thirdCall).to.eql([{
-              changeState: true,
-              counter: 2,
-            }, {
-              state: 10,
-            }]);
-            expect(wrapper.state()).to.eql({ state: 20 });
-          });
-        });
-      });
-
-      it('does not infinitely loop when a PureComponent fires a noop setState in cDU', () => {
-        class Example extends PureComponent {
-          constructor(props) {
-            super(props);
-
-            this.renders = 0;
-            this.state = {
-              a: false,
-              b: false,
-            };
-          }
-
-          componentDidMount() {
-            this.setState({ b: false });
-          }
-
-          componentDidUpdate() {
-            this.setState({ b: false }); // eslint-disable-line react/no-did-update-set-state
-          }
-
-          render() {
-            this.renders += 1;
-            const { a, b } = this.state;
-            return <div>{`${a} ${b} ${this.renders}`}</div>;
-          }
-        }
-
-        const wrapper = shallow(<Example />);
-        expect(wrapper.debug()).to.equal(`<div>
-  false false 1
-</div>`);
-      });
-    });
-
-    describe('Own PureComponent implementation', () => {
-      it('does not update when state and props did not change', () => {
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'init',
-            };
-          }
-
-          shouldComponentUpdate(nextProps, nextState) {
-            return !isEqual(this.props, nextProps) || !isEqual(this.state, nextState);
-          }
-
-          componentDidUpdate() {}
-
-          render() {
-            const { foo } = this.state;
-            return (
-              <div>
-                {foo}
-              </div>
-            );
-          }
-        }
-        const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-        const wrapper = shallow(<Foo id={1} />);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-        wrapper.setState({ foo: 'update' });
-        expect(spy).to.have.property('callCount', 1);
-
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-        wrapper.setProps({ id: 2 });
-        expect(spy).to.have.property('callCount', 2);
-      });
-    });
-
-    describeIf(is('>= 16.3'), 'support getSnapshotBeforeUpdate', () => {
-      it('calls getSnapshotBeforeUpdate and pass snapshot to componentDidUpdate', () => {
-        const spy = sinon.spy();
-        class Foo extends React.Component {
-          constructor(props) {
-            super(props);
-            this.state = {
-              foo: 'bar',
-            };
-          }
-
-          componentDidUpdate(prevProps, prevState, snapshot) {
-            spy('componentDidUpdate', prevProps, this.props, prevState, this.state, snapshot);
-          }
-
-          getSnapshotBeforeUpdate(prevProps, prevState) {
-            spy('getSnapshotBeforeUpdate', prevProps, this.props, prevState, this.state);
-            return { snapshot: 'ok' };
-          }
-
-          render() {
-            spy('render');
-            return <div>foo</div>;
-          }
-        }
-        const wrapper = shallow(<Foo name="foo" />);
-        spy.resetHistory();
-        wrapper.setProps({ name: 'bar' });
-        expect(spy.args).to.deep.equal([
-          ['render'],
-          ['getSnapshotBeforeUpdate', { name: 'foo' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'bar' }],
-          ['componentDidUpdate', { name: 'foo' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'bar' }, { snapshot: 'ok' }],
-        ]);
-        spy.resetHistory();
-        wrapper.setState({ foo: 'baz' });
-        expect(spy.args).to.deep.equal([
-          ['render'],
-          ['getSnapshotBeforeUpdate', { name: 'bar' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'baz' }],
-          ['componentDidUpdate', { name: 'bar' }, { name: 'bar' }, { foo: 'bar' }, { foo: 'baz' }, { snapshot: 'ok' }],
-        ]);
-      });
-    });
   });
 
   it('works with class components that return null', () => {
@@ -4485,134 +2583,6 @@ describe('shallow', () => {
         onClick: divProps.onClick,
       });
       expect(divProps.onClick).to.be.a('function');
-    });
-  });
-
-  describe('lifecycles', () => {
-    it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-        }
-
-        componentDidUpdate() {}
-
-        onChange() {
-          this.setState({ foo: 'onChange update' });
-        }
-
-        render() {
-          const { foo } = this.state;
-          return <div>{foo}</div>;
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-      const wrapper = shallow(<Foo />);
-      wrapper.setState({ foo: 'wrapper setState update' });
-      expect(wrapper.state('foo')).to.equal('wrapper setState update');
-      expect(spy).to.have.property('callCount', 1);
-      wrapper.instance().onChange();
-      expect(wrapper.state('foo')).to.equal('onChange update');
-      expect(spy).to.have.property('callCount', 2);
-    });
-
-    it('calls `componentDidUpdate` when component’s `setState` is called through a bound method', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-          this.onChange = this.onChange.bind(this);
-        }
-
-        componentDidUpdate() {}
-
-        onChange() {
-          // enzyme can't handle the update because `this` is a ReactComponent instance,
-          // not a ShallowWrapper instance.
-          this.setState({ foo: 'onChange update' });
-        }
-
-        render() {
-          const { foo } = this.state;
-          return (
-            <div>
-              {foo}
-              <button type="button" onClick={this.onChange}>click</button>
-            </div>
-          );
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-      const wrapper = shallow(<Foo />);
-      wrapper.find('button').prop('onClick')();
-      expect(wrapper.state('foo')).to.equal('onChange update');
-      expect(spy).to.have.property('callCount', 1);
-    });
-
-    it('calls `componentDidUpdate` when component’s `setState` is called', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-          this.update = () => this.setState({ foo: 'update' });
-        }
-
-        componentDidMount() {
-          this.update();
-        }
-
-        componentDidUpdate() {}
-
-        render() {
-          const { foo } = this.state;
-          return <div>{foo}</div>;
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidUpdate');
-
-      const wrapper = shallow(<Foo />);
-      expect(spy).to.have.property('callCount', 1);
-      expect(wrapper.state('foo')).to.equal('update');
-    });
-
-    it('does not call `componentDidMount` twice when a child component is created', () => {
-      class Foo extends React.Component {
-        constructor(props) {
-          super(props);
-          this.state = {
-            foo: 'init',
-          };
-        }
-
-        componentDidMount() {}
-
-        render() {
-          const { foo } = this.state;
-          return (
-            <div>
-              <button type="button" onClick={() => this.setState({ foo: 'update2' })}>
-                click
-              </button>
-              {foo}
-            </div>
-          );
-        }
-      }
-      const spy = sinon.spy(Foo.prototype, 'componentDidMount');
-
-      const wrapper = shallow(<Foo />);
-      expect(spy).to.have.property('callCount', 1);
-      wrapper.find('button').prop('onClick')();
-      expect(spy).to.have.property('callCount', 1);
     });
   });
 });
