+++ packages/enzyme-test-suite/test/_helpers/version.js
@@ -11,3 +11,7 @@ export function is(range) {
 }
 
 export const REACT16 = is('16');
+
+// The shallow renderer in react 16 does not yet support batched updates. When it does,
+// we should be able to go un-skip all of the tests that are skipped with this flag.
+export const BATCHING = !REACT16;
